ssl-python-clients/requirements.txt
attrs==20.2.0
backcall==0.2.0
decorator==4.4.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
Pygments==2.7.1
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
